pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poker-knight-ng"
version = "0.1.0"
description = "GPU-accelerated Texas Hold'em poker solver"
authors = [{name = "Hildolfr"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["poker", "texas-holdem", "gpu", "cuda", "monte-carlo", "solver"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Games/Entertainment",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: CUDA",
]

dependencies = [
    "cupy-cuda12x>=13.0",
    "numpy>=1.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "mypy",
    "pytest-benchmark",
]

[project.urls]
Homepage = "https://github.com/hildolfr/poker_knightNG"
Repository = "https://github.com/hildolfr/poker_knightNG.git"
Issues = "https://github.com/hildolfr/poker_knightNG/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
poker_knight_ng = ["cuda/*.cu", "cuda/*.cuh"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]